Module serde_xml_rs::de

Structs

Deserializer

Functions

from_reader

A convenience method for deserialize some object from a reader.

from_str

A convenience method for deserialize some object from a string.