Module serde_xml_rs::ser

Structs

Serializer
An XML Serializer.

Functions

to_string
A convenience method for serializing some object to a string.

to_writer
A convenience method for serializing some object to a buffer.