serde_xml_rs
Struct EventReader

pub struct EventReader<R>
where
    R: Read,
{ /* private fields */ }

A wrapper around an std::io::Read instance which provides pull-based XML parsing.

Implementations

impl<R> EventReader<R>
where R: Read,

pub fn new(source: R) -> EventReader<R>

Creates a new reader, consuming the given stream.

pub fn new_with_config( source: R, config: impl Into<ParserConfig2>, ) -> EventReader<R>

Creates a new reader with the provded configuration, consuming the given stream.

pub fn next(&mut self) -> Result<XmlEvent, Error>

Pulls and returns next XML event from the stream.

If this returns Err or XmlEvent::EndDocument then further calls to this method will return this event again.

pub fn skip(&mut self) -> Result<(), Error>

Skips all XML events until the next end tag at the current level.

Convenience function that is useful for the case where you have encountered a start tag that is of no interest and want to skip the entire XML subtree until the corresponding end tag.

pub fn source(&self) -> &R

Access underlying reader

Using it directly while the event reader is parsing is not recommended

pub fn source_mut(&mut self) -> &mut R

Access underlying reader

Using it directly while the event reader is parsing is not recommended

pub fn into_inner(self) -> R

Unwraps this EventReader, returning the underlying reader.

Note that this operation is destructive; unwrapping the reader and wrapping it again with EventReader::new() will create a fresh reader which will attempt to parse an XML document from the beginning.

pub fn doctype(&self) -> Option<&str>

Returns the DOCTYPE of the document if it has already been seen

Available only after the root StartElement event

impl<'r> EventReader<&'r [u8]>

pub fn from_str(source: &'r str) -> EventReader<&'r [u8]>

A convenience method to create an XmlReader from a string slice.